Use a Standard Curve to Calculate Unkown Values

Use a Standard Curve to Calculate Unkown Values

std_curve_predict(model, unknowns, digits = 3)

Arguments

model
A linear model, created with either lm() or standard::std_curve_fit()

unknowns
A numeric vector of unknown values, which the standard curve will be used to predict their values.

Value

a tibble with a column for the unknown values, and a column .fitted for the predicted values, based on the standard curve.

Examples

prot <- c(
  0.000, 0.016, 0.031, 0.063, 0.125, 0.250, 0.500, 1.000,
  0.000, 0.016, 0.031, 0.063, 0.125, 0.250, 0.500, 1.000
)
abs <- c(
  0.329, 0.352, 0.349, 0.379, 0.417, 0.491, 0.668, 0.956,
  0.327, 0.341, 0.355, 0.383, 0.417, 0.446, 0.655, 0.905
)

data <- tibble::tibble(
  abs = abs,
  prot = prot
)

unknowns <- c(0.554, 0.568, 0.705)

data %>%
  standard::std_curve_fit(prot, abs) %>%
  standard::std_curve_predict(unknowns)
#> # A tibble: 3 x 2
#>     abs  prot
#>   <dbl> <dbl>
#> 1 0.554 0.361
#> 2 0.568 0.384
#> 3 0.705 0.609